requirements.txt
python-telegram-bot==20.4
python-dotenv==1.0.0
google-generativeai==0.3.1
flask==2.3.3
